isspace() usage
Replace most uses of isspace() with a simple test for ' ' after
processing the string buffer with mungspaces (which replaces tab
with space, converts instances of consecutive whitespace into a
single space, and removes leading and trailing spaces).  The uses
where this wasn't done now cast their argument to (uchar) so that
platforms with signed chars will never pass negative values to it.

I didn't mess with the menu coloring code (except for casts to the
isspace() argument); it almost certainly could benefit from using
mungspaces.  I did mess with the symset processing quite a bit,
and hope I haven't accidentally broken anything.  Default symbols
and DECgraphics symbols still parse and display ok, so the rest of
dat/symbols should be ok too.  I didn't test symbols in the user's
config file because I don't remember how that's supposed to work.

+++ src/end.c
@@ -1,4 +1,4 @@
-/* NetHack 3.5	end.c	$NHDT-Date: 1425319883 2015/03/02 18:11:23 $  $NHDT-Branch: master $:$NHDT-Revision: 1.81 $ */
+/* NetHack 3.5	end.c	$NHDT-Date: 1429953061 2015/04/25 09:11:01 $  $NHDT-Branch: master $:$NHDT-Revision: 1.93 $ */
 /* NetHack 3.5	end.c	$Date: 2012/04/09 02:56:30 $  $Revision: 1.79 $ */
 /* Copyright (c) Stichting Mathematisch Centrum, Amsterdam, 1985. */
 /* NetHack may be freely redistributed.  See license for details. */
@@ -1605,10 +1605,11 @@ wordcount(p)
     char *p;
 {
     int words = 0;
-    while(*p){
-	while(*p && isspace(*p))p++;
-	if(*p) words++;
-	while(*p && !isspace(*p))p++;
+
+    while (*p) {
+	while (*p && isspace((uchar)*p)) p++;
+	if (*p) words++;
+	while (*p && !isspace((uchar)*p)) p++;
     }
     return words;
 }
@@ -1620,8 +1621,8 @@ bel_copy1(inp, out)
     char *in = *inp;
 
     out += strlen(out);	/* eos() */
-    while(*in && isspace(*in)) in++;
-    while(*in && !isspace(*in)) *out++ = *in++;
+    while (*in && isspace((uchar)*in)) in++;
+    while (*in && !isspace((uchar)*in)) *out++ = *in++;
     *out = '\0';
     *inp = in;
 }
